+++ aliding-20230426/ChangeLog.md
@@ -1,3 +1,11 @@
+2025-08-06 Version: 2.44.0
+- Support API GetCardTemplate.
+- Update API CreateTodoTask: add request parameters remindNotifyConfigs.
+- Update API CreateTodoTask: add request parameters reminderTimeStamp.
+- Update API CreateTodoTask: add request parameters notifyConfigs.sendAssistantChat.
+- Update API CreateTodoTask: add request parameters notifyConfigs.sendTodoApn.
+
+
 2025-07-16 Version: 2.43.0
 - Support API ChangeDingTalkId.